начала делать отправку данных (m11t1)
